AB#1293: Remove Dead Code, Move Helpers for finding DotNet Path to PathHelpers.cs
