+++ src/XMLDoc2Markdown/Utils/AssemblyLoadContext.cs
@@ -1,7 +1,6 @@
 using System;
 using System.IO;
 using System.Reflection;
-using System.Runtime.InteropServices;
 using System.Runtime.Loader;
 
 namespace XMLDoc2Markdown.Utils;
@@ -23,13 +22,12 @@ internal class AssemblyLoadContext : System.Runtime.Loader.AssemblyLoadContext
 
     private Assembly OnResolving(System.Runtime.Loader.AssemblyLoadContext context, AssemblyName name)
     {
-            string[] possiblePaths = new[]
-            {
+            string[] possiblePaths =
+            [
                 Path.GetDirectoryName(this._pluginPath),
-                //@"C:\Program Files\dotnet\shared\Microsoft.AspNetCore.App\8.0.8",
-                GetAspNetCoreSharedPath(),
+                PathHelpers.GetAspNetCoreSharedPath(),
                 AppDomain.CurrentDomain.BaseDirectory
-            };
+            ];
 
             foreach (string searchPath in possiblePaths)
             {
@@ -67,42 +65,5 @@ internal class AssemblyLoadContext : System.Runtime.Loader.AssemblyLoadContext
         }
 
         return IntPtr.Zero;
-    }
-    
-    static string GetDotNetRootPath()
-    {
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            return Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles),"dotnet");
-        }
-
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux) || 
-            RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-        {
-            return "/usr/share/dotnet";
-        }
-
-        throw new PlatformNotSupportedException("Unsupported operating system.");
-    }
-    private static string GetAspNetCoreSharedPath()
-    {
-        string dotnetPath = GetDotNetRootPath();
-        string aspNetPath = Path.Combine(dotnetPath, "shared", "Microsoft.AspNetCore.App");
-        string fullPath = Path.Combine(aspNetPath, GetVersionDirectoryName(aspNetPath));
-        return fullPath;
-    }
-
-    private static string GetVersionDirectoryName(string aspNetPath)
-    {
-        string[] versions = Directory.GetDirectories(aspNetPath, "8.0.*");
-        if (versions.Length == 0)
-        {
-            throw new DirectoryNotFoundException("No directory found starting with 8.0.");
-        }
-
-        Array.Sort(versions);
-        
-        return versions[^1];
- 
     }
 }
